backend/foodgram/requirements.txt
asgiref==3.2.10
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.1.0
flake8==5.0.4
gunicorn==20.0.4
isort==5.10.1
psycopg2-binary==2.8.6
pytz==2020.1
reportlab==3.6.12
sqlparse==0.3.1
